Refactor to move gcam-specific code into a gcam package.
I'm still not entirely happy with the way this is organized. First,
we've overloaded the term "module" to mean a functional unit in the
GCAM workflow, so we have a "modules" module that has the class
definitions for all of those functional units.  Second, these class
definitions have a lot of code that properly belongs in topic-specific
subpackages.  However, because of the way the parser works, we need
to arrange for gcam_driver.py to have all of those names in its
local namespace.  Therefore, I'm going to punt on fixing this up until
I come up with a less clunky way to do this.