util.gcam_query:  Correctly detect bare strings in Python 3
For the file list arguments to gcam_query, if you have only a single file, then
you may pass a string rather than a list with only a single element.  We had
been using the presence or absence of the __iter__ attribute to distinguish
between the two cases, but in Python 3 strings have this attribute because
you can in fact iterate over them.  Therefore, we have to check for the string
instance explicitly.

Theoretically this makes the test for the __iter__ attribute redundant, but
since in python 2 strings are derived from basestr, I left it in to make sure
that we still catch strings in Python 2.