Set RAB status to successful on init.
The RAB will still be running when the master thread goes to check
component exit statuses, so it needs to show a successful status
indicator to avoid throwing a spurious error.  We could add an
exception in the checker loop, but there isn't any value to the RAB
showing a pending status anyhow, so it seems cleanest just to report
success from the moment the system is initialized.